Consolidation
Consolidated all used functionality down into just the .ino file with a utils cpp object.
